Add files via upload
This is an android application software which alerts users with an automatic alarming system for medication and dosages. Most of the time patients cannot remember the exact time when he/she needs to take the medicine. This app will alert patients about the time of medication, description of medicine, before-after meal, and so on. The alarm can be set with different types of pictures with the name of the medicines to overcome any kind of difficulties in reading. To develop this project, agile development strategy is followed to support the customer requirements. Android Studio IDE, Java, and XML tools will be used for the implementation of the project. Basically, the project is an attempt to develop a system which will be time-saving and supportive for medication adherence.